web/requirements.txt
fastapi==0.111.0           # main web framework (ASGI)  👈  FastAPI docs show it pairs with Uvicorn for production use :contentReference[oaicite:0]{index=0}
uvicorn[standard]==0.29.0  # ASGI server with hot-reload & websockets :contentReference[oaicite:1]{index=1}
SQLAlchemy==2.0.29         # SQL toolkit / ORM (async-friendly since v2) :contentReference[oaicite:2]{index=2}
asyncpg==0.29.0            # ultra-fast async PostgreSQL driver :contentReference[oaicite:3]{index=3}
aiogram==3.4.1             # Telegram helpers for login-hash verification :contentReference[oaicite:4]{index=4}
python-jose[cryptography]==3.3.0  # signs the JWT you store in the cookie :contentReference[oaicite:5]{index=5}
jinja2==3.1.3              # HTML templates served by FastAPI :contentReference[oaicite:6]{index=6}
qrcode[pil]==7.4.2
requests==2.32.0
minio==7.2.5
httpx==0.27.0  # async HTTP client for Telegram broadcast
passlib[bcrypt]==1.7.4  # password hashing for email/password auth
alembic==1.13.1  # database migrations
reportlab==4.0.4  # PDF generation for landlord QR bundles
aioredis>=2.0
argon2-cffi>=23.1
slowapi>=0.1.5